+++ src/languageFeatures/internal/diagnostics/shared/diagnosticCodes/relevantWithinSettingsBlockDiagnosticCodeEnum.ts
@@ -1,0 +1,3 @@
+export enum RelevantWithinSettingsBlockDiagnosticCode{
+    EncodeUrlValueInvalid = "bru600",
+}